backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
PyMySQL==1.1.1
cryptography==43.0.0
python-dotenv==1.0.1
Werkzeug==3.0.3
pandas==2.2.2
scikit-learn==1.5.1
requests==2.31.0
